code update

+++ FemtoProtonAnalysis/include/MomentFunctions.h
@@ -4,9 +4,9 @@
 #include <vector>
 #include <utility>
 
-double q_r_s(std::vector<std::pair<int,double>> &track_eff_pair, int r, int s);
+long double q_r_s(std::vector<std::pair<int,double>> &track_eff_pair, int r, int s);
 
-std::vector<std::vector<double>> make_all_q_s( std::vector<std::pair<int,double>> & track_eff_pair_vec, int rMax, int sMax );
+std::vector<std::vector<long double>> make_all_q_s( std::vector<std::pair<int,double>> & track_eff_pair_vec, int rMax, int sMax );
 
 
 #endif
